backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
psycopg[binary]==3.2.1
django-environ==0.11.2
gunicorn==22.0.0
celery==5.4.0
redis==5.0.7
django-cors-headers==4.4.0
python-dateutil==2.9.0.post0
openai>=1.30.0
pydantic>=2.7.0
requests>=2.32.3